Create interface RateLimiter
