+++ agent_api/src/test/java/ratelimiting/FixedWindowRateLimiterTest.java
@@ -1,27 +1,27 @@
 package ratelimiting;
 
-import dev.aikido.agent_api.ratelimiting.RateLimiter;
+import dev.aikido.agent_api.ratelimiting.FixedWindowRateLimiter;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.concurrent.TimeUnit;
 
-class RateLimiterTest {
+class FixedWindowRateLimiterTest {
 
     @Test
     void testAllowUpToMaxAmountRequestsWithinTTL() {
         int maxAmount = 5;
         long ttl = 500; // 0.5 seconds
-        RateLimiter rateLimiter = new RateLimiter(maxAmount, ttl);
+        FixedWindowRateLimiter fixedWindowRateLimiter = new FixedWindowRateLimiter(maxAmount, ttl);
 
         String key = "user1";
         for (int i = 0; i < maxAmount; i++) {
-            assertTrue(rateLimiter.isAllowed(key, ttl, maxAmount),
+            assertTrue(fixedWindowRateLimiter.isAllowed(key, ttl, maxAmount),
                     "Request " + (i + 1) + " should be allowed");
         }
 
-        assertFalse(rateLimiter.isAllowed(key, ttl, maxAmount),
+        assertFalse(fixedWindowRateLimiter.isAllowed(key, ttl, maxAmount),
                 "Request 6 should not be allowed");
     }
 
@@ -29,21 +29,21 @@ class RateLimiterTest {
     void testResetAfterTTLExpired() throws InterruptedException {
         int maxAmount = 5;
         long ttl = 500; // 0.5 seconds
-        RateLimiter rateLimiter = new RateLimiter(maxAmount, ttl);
+        FixedWindowRateLimiter fixedWindowRateLimiter = new FixedWindowRateLimiter(maxAmount, ttl);
 
         String key = "user1";
         for (int i = 0; i < maxAmount; i++) {
-            assertTrue(rateLimiter.isAllowed(key, ttl, maxAmount),
+            assertTrue(fixedWindowRateLimiter.isAllowed(key, ttl, maxAmount),
                     "Request " + (i + 1) + " should be allowed");
         }
 
-        assertFalse(rateLimiter.isAllowed(key, ttl, maxAmount),
+        assertFalse(fixedWindowRateLimiter.isAllowed(key, ttl, maxAmount),
                 "Request 6 should not be allowed");
 
         // Simulate the passage of time
         TimeUnit.MILLISECONDS.sleep(ttl + 100); // Add a small buffer
 
-        assertTrue(rateLimiter.isAllowed(key, ttl, maxAmount),
+        assertTrue(fixedWindowRateLimiter.isAllowed(key, ttl, maxAmount),
                 "Request after TTL should be allowed");
     }
 
@@ -51,25 +51,25 @@ class RateLimiterTest {
     void testAllowRequestsForDifferentKeysIndependently() {
         int maxAmount = 5;
         long ttl = 500; // 0.5 seconds
-        RateLimiter rateLimiter = new RateLimiter(maxAmount, ttl);
+        FixedWindowRateLimiter fixedWindowRateLimiter = new FixedWindowRateLimiter(maxAmount, ttl);
 
         String key1 = "user1";
         String key2 = "user2";
 
         for (int i = 0; i < maxAmount; i++) {
-            assertTrue(rateLimiter.isAllowed(key1, ttl, maxAmount),
+            assertTrue(fixedWindowRateLimiter.isAllowed(key1, ttl, maxAmount),
                     "Request " + (i + 1) + " for key1 should be allowed");
         }
 
-        assertFalse(rateLimiter.isAllowed(key1, ttl, maxAmount),
+        assertFalse(fixedWindowRateLimiter.isAllowed(key1, ttl, maxAmount),
                 "Request 6 for key1 should not be allowed");
 
         for (int i = 0; i < maxAmount; i++) {
-            assertTrue(rateLimiter.isAllowed(key2, ttl, maxAmount),
+            assertTrue(fixedWindowRateLimiter.isAllowed(key2, ttl, maxAmount),
                     "Request " + (i + 1) + " for key2 should be allowed");
         }
 
-        assertFalse(rateLimiter.isAllowed(key2, ttl, maxAmount),
+        assertFalse(fixedWindowRateLimiter.isAllowed(key2, ttl, maxAmount),
                 "Request 6 for key2 should not be allowed");
     }
 }
